Move and Rename specs because minutest seems to prefer _test to _spec
