+++ test/lightblue/client_test.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'test_helper'
 
 describe Lightblue::Client do
   subject do
